+++ PhysicComponent.h
@@ -1,16 +1,15 @@
 #pragma once
 #include "Vec2.h"
 
-// struct definition
+
 struct PhysicComponent
 {
 public:
 	Vec2 Position;
 	Vec2 Velocity;
 	Vec2 Acceleration;
-	//TODO COLLIDER with Circle and distance
 
-	PhysicComponent(float posX, float posY, float velX, float velY);
+	PhysicComponent(double posX, double posY, double velX, double velY);
 
-	float Dist(PhysicComponent* physic);
+	double Dist(PhysicComponent* physic);
 };
